+++ src/com/lvl6/retrieveutils/EquipmentRetrieveUtils.java
@@ -75,7 +75,7 @@ public class EquipmentRetrieveUtils {
     int minLevel = rs.getInt(6);
     int coinPrice = rs.getInt(7);
     int diamondPrice = rs.getInt(8);
-    double chanceOfLoss = rs.getDouble(9);
+    float chanceOfLoss = rs.getFloat(9);
     Equipment equip = null;
     if (coinPrice > 0) {
       equip = new Equipment(id, name, type, attackBoost, defenseBoost, minLevel, coinPrice, Equipment.NOT_SET, chanceOfLoss);
